Preparing to remove the function modulo.
The function modulo in the module Polynome is quite strange. The
function takes two polynomials as arguments and returns a polynomial.
Contrary to its name suggests, the function does not compute a modulo of
two polynomials but the modulo of two rational numbers see as
polynomials. If one of its arguments is not a constant polynomial, the
function raises an exception. Since the function is called only once in
all the code of AE, I decided to erase it completely.

FIX: I fix also a mistake I did during the clean-up of the module
Polynome. I replaced the function find by M.find in the code of the
function add. This modification broke AE!