Modify the division function in module Polynome
The previous div function had the same strange behaviour as the
former function modulo. I replaced it by a new function div_const
that divides a polynomial by a rational number and raises an exception
in case of division by zero.

I also moved a part of the former function in module Arith
because this code is to specific to stay in the module Polynome.