Produce an abstract value for abstract type.
We cannot produce a concrete value for a symbol whose the type is an
abstract value. For instance, consider the input file:

```
(set-logic ALL)
(set-option :produce-models true)
(declare-sort s 0)
(declare-const a s)
(declare-const b s)
(assert (= a b))
(check-sat)
(get-model)
```

Alt-Ergo outputs `unknown`, which makes sense as we don't know anything
about the abstract type `s`. As `a` and `b` are user-defined, we have to
print something in the model about them but we cannot! Thus, we have to
print abstract values for them.