Fix SMT-LIB printers of Expr
This PR fixes the SMT-LIB printers in `Expr` and `Symbols`. As we
discuss offline, we plan to use the type `Expr.t` to represent model
values. To do so, we need to update the printers of `Expr` and `Symbols`
in order to print SMT-LIB compliant terms.

After this commit, the printers `Expr.print` and `Sy.pp_smtlib` should
always print SMT-LIB terms (for `Expr.print`, we need to select the
appropriate output format or using a SMT-LIB input file).

I also clean up the code of the printers by using the library `Fmt` as
much as possible and simplify some cases in the printer SMT-LIB of `Expr`.

As there is no syntax for semantic trigger, I don't print them if we
call the printer on a formula.