Documentation in UF and fix BV symbols
Currently, we produce casesplits for bitvectors in order to generate
models but we cannot reasoning with `BVand`, `BVxor` and `BVor`. It
means we cannot add these symbols as interpreted symbols of the theory
BV (see the function `Bitv.is_mine_symbol`) but these symbols can appear
in the union-find environment. We have to ignore them while computing
the model in UF.