Propagate new terms in SatML
This commit is a huge hack to extra new terms generated by theories in
the deep guts of SatML in order to propagate them to the instantiation
engine. The basic idea of the hack consists in using the exception `Sat`
of SatML and the returned value of `assume` to store these terms.